Move error code constants to own file
