+++ src/Sulu/Bundle/AdminBundle/Resources/js/constants.js
@@ -1,0 +1,3 @@
+// @flow
+export const ERROR_CODE_DEPENDANT_RESOURCES_FOUND = 1105;
+export const ERROR_CODE_REFERENCING_RESOURCES_FOUND = 1106;
